+++ app/src/main/java/com/tcc/printmobile/model/Img.java
@@ -4,7 +4,7 @@ public class Img extends File {
 
 	private static final long serialVersionUID = 1L;
 	
-	public Img(Boolean colorful, Boolean landscape, Long copies) {		
+	public Img(Boolean colorful, Boolean landscape, int copies) {
 		super(colorful, landscape, copies);
 	}
 }
